refactor: 열거형 클래스 패키지 구조 변경
